feat: move shared code from api to generic to be used in translation
