+++ packages/application-generic/src/decorators/external-api.decorator.ts
@@ -1,0 +1,5 @@
+import { SetMetadata } from '@nestjs/common';
+
+// eslint-disable-next-line @typescript-eslint/naming-convention
+export const ExternalApiAccessible = () =>
+  SetMetadata('external_api_accessible', true);
