+++ packages/application-generic/src/decorators/index.ts
@@ -1,0 +1,2 @@
+export * from './user-session.decorator';
+export * from './external-api.decorator';
