+++ packages/application-generic/src/decorators/user-session.decorator.ts
@@ -1,11 +1,8 @@
 import { createParamDecorator, UnauthorizedException } from '@nestjs/common';
 import * as jwt from 'jsonwebtoken';
-import { UserSession } from '@novu/application-generic';
-
-export { UserSession };
 
 // eslint-disable-next-line @typescript-eslint/naming-convention
-export const SubscriberSession = createParamDecorator((data, ctx) => {
+export const UserSession = createParamDecorator((data, ctx) => {
   let req;
   if (ctx.getType() === 'graphql') {
     req = ctx.getArgs()[2].req;
@@ -13,12 +10,22 @@ export const SubscriberSession = createParamDecorator((data, ctx) => {
     req = ctx.switchToHttp().getRequest();
   }
 
-  if (req.user) return req.user;
+  if (req.user) {
+    /**
+     * This helps with sentry and other tools that need to know who the user is based on `id` property.
+     */
+    req.user.id = req.user._id;
+    req.user.username = (req.user.firstName || '').trim();
+    req.user.domain = req.user.email?.split('@')[1];
+
+    return req.user;
+  }
 
   if (req.headers) {
     if (req.headers.authorization) {
       const tokenParts = req.headers.authorization.split(' ');
-      if (tokenParts[0] !== 'Bearer') throw new UnauthorizedException('bad_token');
+      if (tokenParts[0] !== 'Bearer')
+        throw new UnauthorizedException('bad_token');
       if (!tokenParts[1]) throw new UnauthorizedException('bad_token');
 
       const user = jwt.decode(tokenParts[1]);
