+++ packages/application-generic/src/services/auth/index.ts
@@ -1,2 +1,3 @@
 export * from './auth.service';
 export * from './shared';
+export * from './auth.guard';
